Unsigned int indices

+++ js/rendering/IndexData.js
@@ -4,7 +4,7 @@ class IndexData extends RenderingData {
     #count;
 
     constructor(gl, data) {
-        super(gl, gl.ELEMENT_ARRAY_BUFFER, new Uint16Array(data));
+        super(gl, gl.ELEMENT_ARRAY_BUFFER, new Uint32Array(data));
         this.#count = data.length;
     }
